Move serializer package.
