+++ wallet/serializer/serializer_test.go
