+++ wallet/substrates/wallet_wire_processor.go
@@ -4,7 +4,7 @@ import (
 	"errors"
 	"fmt"
 	"github.com/bsv-blockchain/go-sdk/wallet"
-	"github.com/bsv-blockchain/go-sdk/wallet/substrates/serializer"
+	"github.com/bsv-blockchain/go-sdk/wallet/serializer"
 )
 
 // WalletWireProcessor implements the WalletWire interface
